Desktop: Add create/edit inventory methods
There are now methods in the FoodTruckController to create and edit
food supply and equipment counts in the inventory. Each item is
instantiated with a quantity of 0, and it is up to the user to define
the quantity manually. Deleting is part of the edit function as well;
inputting a negative integer will delete the item from the inventory
altogether.