Fix bug making app crash
Fixed a bug making the app crash. What happened is "refreshData" ran "displayShifts" which took as input the element
of the hash map at the index at which the spinner points to. However, when the app is started, the "employee name" spinner
is set to null, so this crashed our app.